refactor: LottoNumberFactory.kt 및 Money.kt 패키기 변경
