+++ src/test/kotlin/lotto/domain/MoneyTest.kt
@@ -1,8 +1,9 @@
-package lotto
+package lotto.domain
 
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatExceptionOfType
 import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.assertAll
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.NullAndEmptySource
 import org.junit.jupiter.params.provider.ValueSource
@@ -15,17 +16,21 @@ class MoneyTest {
         assertThat(Money.of(input).value).isEqualTo(expected)
     }
 
-    @ParameterizedTest
-    @NullAndEmptySource
-    @ValueSource(strings = ["0", "-1000", "abc"])
-    fun `구입 금액이 양의 정수가 아닌 경우 예외를 발생시킨다`(input: String?) {
-        assertThatExceptionOfType(IllegalArgumentException::class.java)
-            .isThrownBy { Money.of(input) }
+    @Test
+    fun `덧셈과 곱셉이 가능하다`() {
+        val money = Money.of(100)
+        val other = Money.of(200)
+        val expected = Money.of(300)
+        assertAll(
+            { assertThat(money + other).isEqualTo(expected) },
+            { assertThat(money * 3).isEqualTo(expected) }
+        )
     }
 
     @ParameterizedTest
-    @ValueSource(strings = ["100", "999"])
-    fun `구입 금액이 1000원보다 낮은 값이 들어올 경우 예외를 발생시킨다`(input: String) {
+    @NullAndEmptySource
+    @ValueSource(strings = ["-1000", "abc"])
+    fun `구입 금액이 양의 정수가 아닌 경우 예외를 발생시킨다`(input: String?) {
         assertThatExceptionOfType(IllegalArgumentException::class.java)
             .isThrownBy { Money.of(input) }
     }
